TrOCR/requirements.txt
transformers==4.41.1
torch>=2.0.0
Pillow
Flask
